% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorb.R
\name{comorb}
\alias{comorb}
\title{Comorbilidades}
\usage{
comorb(
  archivo = "//Centcs01srv03/TABLEAU/Coronavirus/R/PacDiag2021.csv",
  id = "Pac_Unif_Cod",
  code = "Diag_Cod",
  fecha = "ResDiag_Inicio_Fecha",
  assign0 = TRUE,
  mini = NULL
)
}
\arguments{
\item{archivo.}{Archivo refers to the location of the csv file that needs 3 columns. pacUnifCOde (patient identifier),Diag_Cod(ICD9 or ICD10 of the pathology),ResDGP_Fecha_Presta(date of the comorbidity).}

\item{mini.}{This numeric parameters  defines the number of rows from the "archivo" that the package should be passed on.}
}
\value{
modified version of comorbidity to use  ICD9 , ICD10 and CIAP2 to return Charlson score, elixhauser score and first event date.
}
\description{
modified version of comorbidity to work simultaneously with icd9 and icd10, and Smoking group extraction
}
