% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smokerGroup.R
\name{smokerGroup}
\alias{smokerGroup}
\title{Comorbilidades}
\usage{
smokerGroup(
  tabla = "//Centcs01srv03/TABLEAU/Coronavirus/R/Ultimo_DGP_Tabaco.csv",
  Ultimo_DGP_Tabaco.csv,
  exfumador_meses = 12
)
}
\arguments{
\item{tabla.}{Tabla refers to the location of the csv file that needs 3 columns. pacUnifCOde,DGP_Cod=TABACO,ResDGP_Fecha_Presta. Usually corersponds to the extraction of THE LAST AVAILABLE INFORMATION ON SMOKE PER PATIENT.}

\item{exfumador_meses.}{exfumador_meses reffers to the minimum number of months since last time soking to consider one patient as ex_smoker.}
}
\value{
groups whatever data is stored in ATENEA or ANDIA DGP_Cod=TABACO into three smoker groups. Current smoker, ex-smoker, non-smoker
}
\description{
Extraction of smoking group through regular expressions and data of variable.
}
