% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorb.R, R/comorb_previo.R
\name{comorb}
\alias{comorb}
\title{Comorbilidades}
\usage{
comorb(
  tabla = "//Centcs01srv03/TABLEAU/Coronavirus/R/PacDiag2021.csv",
  id = "Pac_Unif_Cod",
  code = "Diag_Cod",
  fecha = "ResDiag_Inicio_Fecha",
  modo = "pesos",
  assign0 = TRUE,
  mini = NULL
)

comorb(
  tabla = "//Centcs01srv03/TABLEAU/Coronavirus/R/PacDiag2021.csv",
  id = "Pac_Unif_Cod",
  code = "Diag_Cod",
  fecha = "ResDiag_Inicio_Fecha",
  modo = "pesos",
  assign0 = TRUE,
  mini = NULL
)
}
\arguments{
\item{archivo.}{Archivo refers to the location of the csv file that needs 3 columns. pacUnifCOde (patient identifier),Diag_Cod(ICD9 or ICD10 of the pathology),ResDGP_Fecha_Presta(date of the comorbidity).}

\item{mini.}{This numeric parameters  defines the number of rows from the "archivo" that the package should be passed on.}

\item{tabla.}{Tabla refers to the location of the csv file that needs 3 columns. pacUnifCOde (patient identifier),Diag_Cod(ICD9 or ICD10 of the pathology),ResDGP_Fecha_Presta(date of the comorbidity).}

\item{modo.}{Determina la salida del paquete. modo= "pesos" es la variable por defecto (42g+score):  de charlson (17 primeros grupos) y elixhauser, con un operador numerico 1/0 segun se hayan dado o no en el periodo estudiado, más la puntuacion de los índices charlson t¡y elixhauser.modo="fechas" (42g+fecha): 42 grupos, con vacios (NA) para todas aquellas comorbilidades que no se dieron y en el resto la fecha de primera aparicion.modo="codigos"S (42g+cie)**:  42 grupos, igualmente vacios si no tuvieron y con un codigo diagnostico (el primero) en caso afirmativo.}
}
\value{
modified version of comorbidity to use  ICD9 , ICD10 and CIAP2 to return Charlson score, elixhauser score and first event date.

modified version of comorbidity to use both ICD9 and ICD10 and generate CHarlson score, elixhauser score or first event date.
}
\description{
modified version of comorbidity to work simultaneously with icd9 and icd10, and Smoking group extraction

modified version of comorbidity to work simultaneously with icd9 and icd10, and Smoking group extraction
}
\examples{
comorb(archivo="//Centcs01srv03/TABLEAU/Coronavirus/R/PacDiag2021.csv",id="Pac_Unif_Cod",code="Diag_Cod", fecha="ResDiag_Inicio_Fecha",assign0 = TRUE)


comorb(tabla="//Centcs01srv03/TABLEAU/Coronavirus/R/PacDiag2021.csv",id="Pac_Unif_Cod",code="Diag_Cod", fecha="ResDiag_Inicio_Fecha", modo="pesos",assign0 = TRUE)


}
