% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorb.R
\name{comorb}
\alias{comorb}
\title{Comorbilidades}
\usage{
comorb(
  tabla = "//Centcs01srv03/TABLEAU/Coronavirus/R/Nuevos_positivos.csv",
  id = "Pac_Unif_Cod",
  code = "Diag_Cod",
  score = "charlson",
  assign0 = TRUE
)
}
\arguments{
\item{tabla}{}
}
\value{
modified version of comorbidity
}
\description{
modified version of comorbidity to work simultaneously with icd9 and icd10
}
\examples{


}
